+++ synapse/tests/test_cortex.py
@@ -1,9 +1,7 @@
 import time
 import asyncio
 
-
 import synapse.exc as s_exc
-import synapse.glob as s_glob
 import synapse.common as s_common
 import synapse.telepath as s_telepath
 import synapse.datamodel as s_datamodel
@@ -132,7 +130,6 @@ class CortexTest(s_t_utils.SynTest):
             self.len(1, await alist(core.eval('test:str~="zip"')))
             self.len(1, await alist(core.eval('.favcolor~="^r"')))
 
-    @s_glob.synchelp
     async def test_feed_conf(self):
 
         async with self.getTestCryo() as cryo: