+++ synapse/tests/test_lib_layer.py
@@ -4,7 +4,7 @@ from unittest.mock import patch
 import synapse.exc as s_exc
 import synapse.cortex as s_cortex
 import synapse.tests.utils as s_t_utils
-import synapse.lib.layer as s_layer
+from synapse.tests.utils import alist
 
 
 async def iterPropForm(self, form=None, prop=None):
@@ -51,6 +51,52 @@ class LayerTest(s_t_utils.SynTest):
                     await liftByHandler('prop:ival', 1)
                     await liftByHandler('univ:ival', 1)
                     await liftByHandler('form:ival', 1)
+
+    async def test_layer_buidcache(self):
+        '''
+        Make sure the layer buidcache isn't caching incorrectly
+        '''
+
+        async with self.getTestCore() as core:
+            buidcache = core.view.layers[0].buidcache
+
+            async with await core.snap() as snap:
+                node = await snap.addNode('test:str', 'a', {'tick': '2001'})
+                await node.addTag('foo')
+                testbuid = node.buid
+
+            self.isin(testbuid, buidcache)
+            propbag = buidcache.get(testbuid)
+            self.eq(propbag.get('*test:str'), 'a')
+            self.eq(propbag.get('tick'), 978307200000)
+            self.eq(propbag.get('#foo'), (None, None))
+            self.isin('.created', propbag)
+
+            async with await core.snap() as snap:
+                nodes = await alist(snap.getNodesBy('test:str'))
+                self.len(1, nodes)
+                node = nodes[0]
+                self.eq(node.props['tick'], 978307200000)
+                await node.set('tick', '2002')
+
+            self.eq(buidcache.get(testbuid).get('tick'), 1009843200000)
+
+            # new snap -> no cached buids in snap
+            async with await core.snap() as snap:
+                nodes = await alist(snap.getNodesBy('test:str'))
+                self.len(1, nodes)
+                node = nodes[0]
+                self.eq(node.props['tick'], 1009843200000,)
+                self.eq(node.tags, {'foo': (None, None)})
+                await node.delete()
+
+            self.notin(testbuid, buidcache)
+
+            async with await core.snap() as snap:
+                nodes = await alist(snap.getNodesBy('test:str'))
+                self.len(0, nodes)
+
+            self.notin(testbuid, buidcache)
 
     async def test_splicemigration_pre010(self):
         async with self.getRegrCore('pre-010') as core: